+++ public/vocables/index.html
@@ -111,6 +111,7 @@
         <script type="module" src="modules/tools/KGcreator/KGcreator_mappings.js"></script>
         <script type="module" src="modules/tools/KGcreator/KGcreator_graph.js"></script>
         <script type="module" src="modules/tools/KGcreator/KGcreator_run.js"></script>
+        <script type="module" src="modules/tools/KGcreator/mappingModeler.js"></script>
         <script type="module" src="modules/tools/KGcreator/KGcreator_joinTables.js"></script>
 
         <!--   <script type="module" src="modules/tools/KGcreatorGraph.js"></script>-->
@@ -121,7 +122,7 @@
 
         <script type="module" src="modules/shared/authentification.js"></script>
 
-        <script type="module" src="modules/tools/axioms/axiom_editorUI.js"></script>
+        <script type="module" src="modules/tools/axioms/html/old/axiom_editorUI.js"></script>
         <script type="module" src="modules/tools/axioms/axiom_editor.js"></script>
         <script type="module" src="modules/tools/axioms/axioms_graph.js"></script>
         <script type="module" src="modules/tools/axioms/axioms_manager.js"></script>
