+++ public/vocables/modules/tools/KGcreator/KGcreator.js
@@ -89,17 +89,7 @@ var KGcreator = (function () {
                     UI.resetWindowHeight();
                     $("#KGcreator_dialogDiv").dialog({
                         autoOpen: false,
-                        /*   close: function (event, ui) {
-                            window.scrollTo(0, 0);
-                        },
-                        drag: function (event, ui) {
-                            $("#KGcreator_dialogDiv").parent().css("transform", "unset");
-                        },
-                        open(event, ui) {
-                            $("#KGcreator_dialogDiv").parent().css("transform", "translate(-50%,-50%)");
-                            $("#KGcreator_dialogDiv").parent().css("top", "50%");
-                            $("#KGcreator_dialogDiv").parent().css("left", "50%");
-                        },*/
+
                     });
                 });
             });
