+++ public/vocables/modules/tools/KGcreator/html/mappingModeler.html
@@ -1,0 +1,52 @@
+<style></style>
+
+<div style="display: flex; flex-direction: column; min-height: 650px">
+    <div style="width: 100%; height: 25%; overflow: auto">
+        <div>
+            <div id="mappingModeler_infosDiv" style="color: #278ecc"></div>
+            <div id="mappingModeler_axiomsJstreeDiv"></div>
+        </div>
+        <!--
+    <div style="display: flex  ;flex-direction: column;width:100%">
+    <div style="display: flex; flex-direction: row; width: 100%; height: 100px; margin: 5px; padding: 5px">
+        <div id="mappingModeler_axiomText"></div>
+    </div>
+    </div>-->
+        <div style="display: flex; flex-direction: row; width: 100%">
+            <div>
+                <div style="display: flex; flex-direction: column; display: none" id="mappingModeler_graphPanelDiv">
+                    <!--    <button class="w3-button slsv-button-3" onclick="mappingModeler.actions.toGraphMl()">toGraphML
+                    </button>
+                    <button class="w3-button slsv-button-3" onclick="mappingModeler.actions.toSVG()">toSVG</button>-->
+                    <button class="w3-button slsv-button-3" onclick="mappingModeler.actions.showTriples()">show triples</button>
+                    <hr />
+                </div>
+
+                <div style="display: flex; flex-direction: column">
+                    <!--   <button class="w3-button slsv-button-2" onclick="mappingModeler.newAxiom()">New axiom</button>-->
+                    <div id="mappingModeler_activeLegendDiv" style="width: 100px; height: 200px; background-color: #ddd"></div>
+                </div>
+            </div>
+            <div style="display: flex; flex-direction: row; display: none" id="mappingModeler_newAxiomPanel">
+                <!--  <button onclick="mappingModeler.actions.setLayout('randomlayout')">Random Layout</button>-->
+
+                <div>
+                    New Axiom<br />
+                    <select id="axioms_legend_suggestionsSelect" onclick="Axiom_activeLegend.onSuggestionsSelect($(this).val())" size="20" style="width: 250px; overflow: auto"></select>
+                    <div>
+                        <button class="w3-button slsv-button-3" onclick="Axiom_activeLegend.clearAxiom()">clear</button>
+                        <button class="w3-button slsv-button-3" onclick="Axiom_activeLegend.showTriples()">triples</button>
+                        <button class="w3-button slsv-button-3" onclick="Axiom_activeLegend.saveAxiom()">save</button>
+                    </div>
+                </div>
+            </div>
+
+            <div id="mappingModeler_graphDiv" style="width: 70vw; height: 450px; background-color: #ddd; margin: 5px; padding: 5px; overflow: auto"></div>
+        </div>
+    </div>
+
+    <div id="mappingModeler_manchesterDiv" style="display: none">
+        <button class="w3-button slsv-button-3" onclick="Axiom_activeLegend.axiomTriplesToManchester()">toManchester</button>
+        <span id="axiomsEditor_textDiv" style="width: 80%; height: 50px; background-color: #ddd"></span>
+    </div>
+</div>
