+++ public/vocables/modules/tools/KGcreator/mappingModeler.js
@@ -1,0 +1,17 @@
+
+
+var MappingModeler = (function () {
+    var self = {}
+    self.init = function (source, resource, divId) {
+        $("#" + divId).load("modules/tools/KGcreator/html/mappingModeler.html", function () {
+            $("#" + divId).dialog("open");
+            M.drawLegend("nodeInfosAxioms_activeLegendDiv");
+
+        })
+    }
+    return self;
+})()
+
+
+export default MappingModeler;
+window.MappingModeler = MappingModeler;
