+++ public/vocables/modules/tools/axioms/html/old/leftPanel.html
