+++ public/vocables/modules/tools/axioms/html/old/mainPanel.html
