+++ public/vocables/modules/tools/axioms/html/old/mainPanel2.html
